+++ internal/provider/provider.go
@@ -120,6 +120,9 @@ func (p *MetabaseProvider) DataSources(ctx context.Context) []func() datasource.
 			return &CurrentUserDataSource{provider: p}
 		},
 		func() datasource.DataSource {
+			return &DatabaseDataSource{provider: p}
+		},
+		func() datasource.DataSource {
 			return &PermissionsGroupDataSource{provider: p}
 		},
 		func() datasource.DataSource {