+++ internal/schema/database_test.go
@@ -2,7 +2,8 @@ package schema
 
 import (
 	"context"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
+	dSchema "github.com/hashicorp/terraform-plugin-framework/datasource/schema"
+	rSchema "github.com/hashicorp/terraform-plugin-framework/resource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/int64planmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/listplanmodifier"
 	"github.com/stretchr/testify/assert"
@@ -15,16 +16,16 @@ func TestDatabaseResource(t *testing.T) {
 
 	ctx := context.Background()
 
-	t.Run("database schema should return expected fields", func(t *testing.T) {
+	t.Run("database rSchema should return expected fields", func(t *testing.T) {
 		resourceSchema := DatabaseResource()
 
 		assert.NotEmpty(t, resourceSchema.Description)
 		assert.Equal(t, 7, len(resourceSchema.Attributes))
 
 		t.Run("id should be configured", func(t *testing.T) {
-			assert.IsType(t, schema.Int64Attribute{}, resourceSchema.Attributes["id"])
+			assert.IsType(t, rSchema.Int64Attribute{}, resourceSchema.Attributes["id"])
 
-			id := resourceSchema.Attributes["id"].(schema.Int64Attribute)
+			id := resourceSchema.Attributes["id"].(rSchema.Int64Attribute)
 			assert.NotEmpty(t, id.Description)
 			assert.True(t, id.IsComputed())
 			assert.False(t, id.IsRequired())
@@ -33,9 +34,9 @@ func TestDatabaseResource(t *testing.T) {
 		})
 
 		t.Run("engine should be configured", func(t *testing.T) {
-			assert.IsType(t, schema.StringAttribute{}, resourceSchema.Attributes["engine"])
+			assert.IsType(t, rSchema.StringAttribute{}, resourceSchema.Attributes["engine"])
 
-			engine := resourceSchema.Attributes["engine"].(schema.StringAttribute)
+			engine := resourceSchema.Attributes["engine"].(rSchema.StringAttribute)
 			assert.NotEmpty(t, engine.Description)
 			assert.True(t, engine.IsRequired())
 			assert.Contains(t, engine.StringValidators(), validators.IsKnownDatabaseEngineValidator())
@@ -43,17 +44,17 @@ func TestDatabaseResource(t *testing.T) {
 		})
 
 		t.Run("name should be configured", func(t *testing.T) {
-			assert.IsType(t, schema.StringAttribute{}, resourceSchema.Attributes["name"])
+			assert.IsType(t, rSchema.StringAttribute{}, resourceSchema.Attributes["name"])
 
-			name := resourceSchema.Attributes["name"].(schema.StringAttribute)
+			name := resourceSchema.Attributes["name"].(rSchema.StringAttribute)
 			assert.NotEmpty(t, name.Description)
 			assert.True(t, name.IsRequired())
 		})
 
 		t.Run("features should be configured", func(t *testing.T) {
-			assert.IsType(t, schema.ListAttribute{}, resourceSchema.Attributes["features"])
+			assert.IsType(t, rSchema.ListAttribute{}, resourceSchema.Attributes["features"])
 
-			features := resourceSchema.Attributes["features"].(schema.ListAttribute)
+			features := resourceSchema.Attributes["features"].(rSchema.ListAttribute)
 			assert.NotEmpty(t, features.Description)
 			assert.True(t, features.IsComputed())
 			assert.False(t, features.IsRequired())
@@ -62,30 +63,89 @@ func TestDatabaseResource(t *testing.T) {
 		})
 
 		t.Run("details should be configured", func(t *testing.T) {
-			assert.IsType(t, schema.StringAttribute{}, resourceSchema.Attributes["details"])
+			assert.IsType(t, rSchema.StringAttribute{}, resourceSchema.Attributes["details"])
 
-			details := resourceSchema.Attributes["details"].(schema.StringAttribute)
+			details := resourceSchema.Attributes["details"].(rSchema.StringAttribute)
 			assert.NotEmpty(t, details.Description)
 			assert.True(t, details.IsOptional())
 		})
 
 		t.Run("details_secure should be configured", func(t *testing.T) {
-			assert.IsType(t, schema.StringAttribute{}, resourceSchema.Attributes["details_secure"])
+			assert.IsType(t, rSchema.StringAttribute{}, resourceSchema.Attributes["details_secure"])
 
-			detailsSecure := resourceSchema.Attributes["details_secure"].(schema.StringAttribute)
+			detailsSecure := resourceSchema.Attributes["details_secure"].(rSchema.StringAttribute)
 			assert.NotEmpty(t, detailsSecure.Description)
 			assert.True(t, detailsSecure.IsOptional())
 			assert.True(t, detailsSecure.IsSensitive())
 		})
 
 		t.Run("schedules should be configured", func(t *testing.T) {
-			assert.IsType(t, schema.MapAttribute{}, resourceSchema.Attributes["schedules"])
+			assert.IsType(t, rSchema.MapAttribute{}, resourceSchema.Attributes["schedules"])
 
-			schedules := resourceSchema.Attributes["schedules"].(schema.MapAttribute)
+			schedules := resourceSchema.Attributes["schedules"].(rSchema.MapAttribute)
 			assert.NotEmpty(t, schedules.Description)
 			assert.True(t, schedules.IsComputed())
 			assert.False(t, schedules.IsRequired())
 			assert.False(t, schedules.IsOptional())
+		})
+	})
+}
+
+func TestDatabaseDataSource(t *testing.T) {
+	t.Parallel()
+
+	t.Run("database rSchema should return expected fields", func(t *testing.T) {
+		dataSourceSchema := DatabaseDataSource()
+
+		assert.NotEmpty(t, dataSourceSchema.Description)
+		assert.Equal(t, 6, len(dataSourceSchema.Attributes))
+
+		t.Run("id should be configured", func(t *testing.T) {
+			assert.IsType(t, dSchema.Int64Attribute{}, dataSourceSchema.Attributes["id"])
+
+			id := dataSourceSchema.Attributes["id"].(dSchema.Int64Attribute)
+			assert.NotEmpty(t, id.Description)
+			assert.True(t, id.IsRequired())
+		})
+
+		t.Run("engine should be configured", func(t *testing.T) {
+			assert.IsType(t, dSchema.StringAttribute{}, dataSourceSchema.Attributes["engine"])
+
+			engine := dataSourceSchema.Attributes["engine"].(dSchema.StringAttribute)
+			assert.NotEmpty(t, engine.Description)
+			assert.True(t, engine.IsComputed())
+		})
+
+		t.Run("name should be configured", func(t *testing.T) {
+			assert.IsType(t, dSchema.StringAttribute{}, dataSourceSchema.Attributes["name"])
+
+			name := dataSourceSchema.Attributes["name"].(dSchema.StringAttribute)
+			assert.NotEmpty(t, name.Description)
+			assert.True(t, name.IsComputed())
+		})
+
+		t.Run("features should be configured", func(t *testing.T) {
+			assert.IsType(t, dSchema.ListAttribute{}, dataSourceSchema.Attributes["features"])
+
+			features := dataSourceSchema.Attributes["features"].(dSchema.ListAttribute)
+			assert.NotEmpty(t, features.Description)
+			assert.True(t, features.IsComputed())
+		})
+
+		t.Run("details should be configured", func(t *testing.T) {
+			assert.IsType(t, dSchema.StringAttribute{}, dataSourceSchema.Attributes["details"])
+
+			details := dataSourceSchema.Attributes["details"].(dSchema.StringAttribute)
+			assert.NotEmpty(t, details.Description)
+			assert.True(t, details.IsComputed())
+		})
+
+		t.Run("schedules should be configured", func(t *testing.T) {
+			assert.IsType(t, dSchema.MapAttribute{}, dataSourceSchema.Attributes["schedules"])
+
+			schedules := dataSourceSchema.Attributes["schedules"].(dSchema.MapAttribute)
+			assert.NotEmpty(t, schedules.Description)
+			assert.True(t, schedules.IsComputed())
 		})
 	})
 }